Convert to FP32

+++ algorithms/solarArrayReference/solarArrayReferenceAlgorithm.h
@@ -16,27 +16,27 @@ enum attitudeFrame { referenceFrame = 0, bodyFrame = 1 };
 class SolarArrayReferenceAlgorithm final {
    public:
     void reset();
-    SolarArrayReferenceOutput update(const Eigen::Vector3d& sigma_BN,
-                                     const Eigen::Vector3d& sigma_RN,
-                                     const Eigen::Vector3d& vehSunPntBdy,
-                                     double theta,
+    SolarArrayReferenceOutput update(const Eigen::Vector3f& sigma_BN,
+                                     const Eigen::Vector3f& sigma_RN,
+                                     const Eigen::Vector3f& vehSunPntBdy,
+                                     float theta,
                                      uint64_t callTime);
 
-    void setA1Hat_B(const Eigen::Vector3d& axis);
-    Eigen::Vector3d getA1Hat_B() const;
-    void setA2Hat_B(const Eigen::Vector3d& normal);
-    Eigen::Vector3d getA2Hat_B() const;
+    void setA1Hat_B(const Eigen::Vector3f& axis);
+    Eigen::Vector3f getA1Hat_B() const;
+    void setA2Hat_B(const Eigen::Vector3f& normal);
+    Eigen::Vector3f getA2Hat_B() const;
     void setAttitudeFrame(int frame);
     int getAttitudeFrame() const;
 
    private:
-    Eigen::Vector3d a1Hat_B{Eigen::Vector3d::Zero()};  //!< solar array drive axis in body frame coordinates
-    Eigen::Vector3d a2Hat_B{Eigen::Vector3d::Zero()};  //!< solar array surface normal at zero rotation
+    Eigen::Vector3f a1Hat_B{Eigen::Vector3f::Zero()};  //!< solar array drive axis in body frame coordinates
+    Eigen::Vector3f a2Hat_B{Eigen::Vector3f::Zero()};  //!< solar array surface normal at zero rotation
     int attitudeFrame{};  //!< flag = 1: compute theta reference based on current attitude instead of attitude reference
 
-    int count{};           //!< counter variable for finite differences
-    uint64_t priorT{};     //!< prior call time for finite differences
-    double priorThetaR{};  //!< prior output msg for finite differences
+    int count{};          //!< counter variable for finite differences
+    uint64_t priorT{};    //!< prior call time for finite differences
+    float priorThetaR{};  //!< prior output msg for finite differences
 };
 
 #endif  // F32XMERA_SOLAR_ARRAY_REFERENCE_ALGORITHM_H
