+++ algorithms/solarArrayReference/solarArrayReferenceTypes.h
@@ -3,8 +3,8 @@
 
 /*! @brief Output of the solar array reference algorithm. */
 struct SolarArrayReferenceOutput {
-    double theta{};     //!< [rad] reference rotation angle
-    double thetaDot{};  //!< [rad/s] reference rotation angle rate
+    float theta{};     //!< [rad] reference rotation angle
+    float thetaDot{};  //!< [rad/s] reference rotation angle rate
 };
 
 #endif  // F32XMERA_SOLAR_ARRAY_REFERENCE_TYPES_H
